Module anima_engine::math

A mod containing linear math constructs useful in graphics.

Structs

Bezier
A struct useful for creating square and cubic Bézier curves.

BezierPath
A struct useful for creating a path of Bézier curves.

Interpolator
A struct useful to easily compute interpolation ratios.

Matrix
A simple matrix struct tailored specifically for graphics.

Quaternion
A simple quaterion struct tailored specifically for graphics.

Vector
A simple vector struct tailored specifically for graphics.

Enums

Behavior
An enum containing useful interpolation techniques.

Traits

Interpolate
A trait to be implemented by structures that need to handle 2-way interpolation. (not necessarily linear)